finish 5.19
something interesting i observed and wanted to write down:
I wanted to create a new blog just to see exactly how it would display on the screen
but i kept getting error 500 in the console. the backend was throwing an error that "user didn't have _id property"
after some debugging and some console.logs, i noticed that the id stored in the db and the one i was seeing on the backend console were different
then i realized that every time i ran the test, a new 'root' user was created, thus the id would be different, but i never 'logged in' in the frontend using that 'new user'.
so the frontend was passing on old id to the request, and when querying the db an error would be thrown that the id didn't exist...